NLP/SIRI_BIA/API/requirements.txt
fastapi
uvicorn
qdrant-client
openai
llama-index
python-dotenv
llama-index-vector-stores-qdrant
pydantic